// src/types/types.ts
export interface Article {
  id: number;
  featured: boolean;
  title: string;
  url: string;
  imageUrl: string;
  newsSite: string;
  summary: string;
  publishedAt: string;
  launches: [
    {
      id: string;
      provider: string;
    }
  ];
  events: [
    {
      id: string;
      provider: string;
    }
  ];
}

export interface ArticlesState {
  status: string;
  article: Article | null;
  articles: Article[];
  error: string | null;
}